+++ packages/tools/utils/src/string-[string-string].spec.js
@@ -1,13 +1,20 @@
 import {strict as assert} from 'assert';
 
-import {prepend} from './string-[string-string]';
+import {makePostfixed, makePrefixed} from './string-[string-string]';
 
 describe('String -> (String -> String)', function() {
-	describe('prepend', function() {
+	describe('makePostfixed', function() {
+		it('should return a function that appends the provided string to the input string', function() {
+			const postfixed = makePostfixed('---');
+			assert.deepStrictEqual(postfixed('A'), 'A---');
+			assert.deepStrictEqual(postfixed('B'), 'B---');
+		});
+	});
+	describe('makePrefixed', function() {
 		it('should return a function that prepends the provided string to the input string', function() {
-			const prependFoo = prepend('foo');
-			assert.deepStrictEqual(prependFoo('A'), 'fooA');
-			assert.deepStrictEqual(prependFoo('B'), 'fooB');
+			const prefixed = makePrefixed('---');
+			assert.deepStrictEqual(prefixed('A'), '---A');
+			assert.deepStrictEqual(prefixed('B'), '---B');
 		});
 	});
 });